Convert to int

+++ src/lambda/Payment/lambda_function.py
@@ -15,7 +15,7 @@ def handler(event, context):
     try:
         req_body = json.loads(event["body"])
         if "payment_method_id" in req_body:
-            _amount = float(req_body["amount"]) * 100
+            _amount = int(float(req_body["amount"]) * 100)
             intent = stripe.PaymentIntent.create(
                 payment_method = req_body["payment_method_id"],
                 amount = _amount,
